+++ src/layers/__tests__/Layer.spec.js
@@ -1,17 +1,29 @@
-window.URL.createObjectURL = jest.fn() // Not supported by jsdom
-
 import Layer from '../Layer'
-// import { addImages } from '../../utils/images'
-jest.mock('../../utils/images')
 
-const mockMapGL = {
-    addLayer: jest.fn(),
-    addSource: jest.fn(),
-    getLayer: jest.fn(),
-}
-const mockMap = {
-    getMapGL: () => mockMapGL,
-}
+const data = [
+    {
+        type: 'Feature',
+        properties: {
+            id: 'O6uvpzGd5pu',
+            name: 'Bo',
+        },
+        geometry: {
+            type: 'Polygon',
+            coordinates: [],
+        },
+    },
+    {
+        type: 'Feature',
+        properties: {
+            id: 'fdc6uOvgoji',
+            name: 'Bombali',
+        },
+        geometry: {
+            type: 'Polygon',
+            coordinates: [],
+        },
+    },
+]
 
 describe('Layer', () => {
     beforeEach(() => {
@@ -31,8 +43,6 @@ describe('Layer', () => {
         expect(layer.getMap()).toBe(mockMap)
         expect(layer.getMapGL()).toBe(mockMapGL)
         mockMapGL.getLayer.mockReturnValueOnce(true)
-        // expect(layer.isOnMap()).toBe(true)
-        // expect(mockMapGL.getLayer).toHaveBeenCalledWith(layer._id)
     })
     it('Should add a non-interactive layer', () => {
         const layer = new Layer()
@@ -58,12 +68,31 @@ describe('Layer', () => {
         expect(layer.getLayers()).toHaveLength(1)
         expect(layer.getLayers()[0].id).toBe(mockMapLayer.id)
         expect(layer.hasLayerId(42)).toBe(true)
-
-        // expect(layer.isOnMap()).toBe(true)
         expect(layer.isVisible()).toBe(true)
+    })
+    it('Should add data features and create numeric ids', () => {
+        const layer = new Layer({ data })
+        const features = layer.getFeatures()
 
-        // expect(layer.isInteractive()).toBe(true)
-        // expect(layer.getInteractiveIds()).toHaveLength(1)
-        // expect(layer.getInteractiveIds()[0]).toBe(mockMapLayer.id)
+        expect(features.length).toBe(data.length)
+        expect(features.every(f => typeof f.id === 'number')).toBe(true)
+        expect(layer.getFeature(2)).toBe(layer.getFeature('fdc6uOvgoji'))
+        expect(layer.getFeature(3)).toBe(undefined)
+    })
+    it('Should set feature hover state', () => {
+        const layer = new Layer({ data })
+        const mockFn = mockMap.setHoverState
+        const source = layer.getId()
+
+        layer.addTo(mockMap)
+        layer.highlight('fdc6uOvgoji')
+        expect(mockFn).toHaveBeenCalled()
+        expect(mockFn).lastCalledWith({ id: 2, source })
+        layer.highlight(1)
+        expect(mockFn).toHaveBeenCalledTimes(2)
+        expect(mockFn).lastCalledWith({ id: 1, source })
+        layer.highlight('abc')
+        expect(mockFn).toHaveBeenCalledTimes(3)
+        expect(mockFn).lastCalledWith(null)
     })
 })